restructured folder structure of passes
