+++ src/qrisp/circuit/pass_management/__init__.py
@@ -16,15 +16,8 @@
 ********************************************************************************
 """
 
-from qrisp.circuit.qubit import *
-from qrisp.circuit.clbit import *
-from qrisp.circuit.operation import *
-from qrisp.circuit.instruction import *
-from qrisp.circuit.quantum_circuit import *
-from qrisp.circuit.compilation_acceleration import *
-from qrisp.circuit.transpiler import *
-from qrisp.circuit.standard_operations import *
-from qrisp.circuit.controlled_operations import *
 from qrisp.circuit.pass_management.circuit_pass import CircuitPass
 from qrisp.circuit.pass_management.pass_manager import PassManager
 from qrisp.circuit.pass_management.passes import *
+
+__all__ = ["CircuitPass", "PassManager"]
