+++ src/qrisp/circuit/pass_management/passes/__init__.py
@@ -16,8 +16,6 @@
 ********************************************************************************
 """
 
-from qrisp.circuit.pass_management.circuit_pass import CircuitPass
-from qrisp.circuit.pass_management.pass_manager import PassManager
 from qrisp.circuit.pass_management.passes.arange_swaps import arange_swaps
 from qrisp.circuit.pass_management.passes.cancel_inverses import cancel_inverses
 from qrisp.circuit.pass_management.passes.cancel_zero_controls import cancel_zero_controls
@@ -25,9 +23,22 @@ from qrisp.circuit.pass_management.passes.combine_single_qubit_gates import comb
 from qrisp.circuit.pass_management.passes.commute_swaps import commute_swaps
 from qrisp.circuit.pass_management.passes.convert_to_cz import convert_to_cz
 from qrisp.circuit.pass_management.passes.gray_synth_toffoli import gray_synth_toffoli, is_toffoli
+from qrisp.circuit.pass_management.passes.manual_layout import manual_layout
 from qrisp.circuit.pass_management.passes.remove_barriers import remove_barriers
 from qrisp.circuit.pass_management.passes.resolve_swaps import resolve_swaps
 from qrisp.circuit.pass_management.passes.reverse_parallelize import reverse_parallelize
-from qrisp.circuit.pass_management.passes.manual_layout import manual_layout
 
-__all__ = ["CircuitPass", "PassManager", "arange_swaps", "cancel_inverses", "cancel_zero_controls", "combine_single_qubit_gates", "commute_swaps", "convert_to_cz", "gray_synth_toffoli", "is_toffoli", "manual_layout", "remove_barriers", "resolve_swaps", "reverse_parallelize"]
+__all__ = [
+    "arange_swaps",
+    "cancel_inverses",
+    "cancel_zero_controls",
+    "combine_single_qubit_gates",
+    "commute_swaps",
+    "convert_to_cz",
+    "gray_synth_toffoli",
+    "is_toffoli",
+    "manual_layout",
+    "remove_barriers",
+    "resolve_swaps",
+    "reverse_parallelize",
+]
