+++ src/qrisp/circuit/pass_management/passes/commute_swaps.py
@@ -21,7 +21,7 @@ from __future__ import annotations
 from qrisp.circuit.quantum_circuit import QuantumCircuit
 from qrisp.circuit.operation import Operation
 from qrisp.circuit.qubit import Qubit
-from qrisp.circuit.passes.pass_manager import CircuitPass
+from qrisp.circuit.pass_management.circuit_pass import CircuitPass
 
 
 @CircuitPass
