+++ src/qrisp/circuit/pass_management/passes/convert_to_cz.py
@@ -21,7 +21,7 @@ from __future__ import annotations
 from collections.abc import Callable
 
 from qrisp.circuit.quantum_circuit import QuantumCircuit
-from qrisp.circuit.passes.pass_manager import CircuitPass
+from qrisp.circuit.pass_management.circuit_pass import CircuitPass
 
 
 def convert_to_cz(
