+++ tests/circuit_tests/passes/test_pass_manager.py
@@ -19,7 +19,7 @@
 import pytest
 from qrisp import PassManager, CircuitPass
 from qrisp.circuit import QuantumCircuit
-from qrisp.circuit.passes import PassManager as PassManagerFromCircuit
+from qrisp.circuit.pass_management import PassManager as PassManagerFromCircuit
 
 
 # ---------------------------------------------------------------------------
@@ -50,12 +50,12 @@ class TestPassManagerImports:
         assert PM is PassManager
 
     def test_import_from_qrisp_circuit_passes(self):
-        """PassManager must be importable from qrisp.circuit.passes."""
+        """PassManager must be importable from qrisp.circuit.pass_management."""
         assert PassManagerFromCircuit is PassManager
 
     def test_import_from_qrisp_circuit_passes_pass_manager(self):
         """PassManager must be importable from the leaf module."""
-        from qrisp.circuit.passes.pass_manager import PassManager as PM
+        from qrisp.circuit.pass_management.pass_manager import PassManager as PM
         assert PM is PassManager
 
 
@@ -254,7 +254,7 @@ class TestPassManagerVerify:
     # --- Unitary verification ---
 
     def test_unitary_all_pass(self):
-        from qrisp.circuit.passes.cancel_inverses import cancel_inverses
+        from qrisp.circuit.pass_management.passes.cancel_inverses import cancel_inverses
 
         pm = PassManager([cancel_inverses, identity_pass])
         qc = QuantumCircuit(2)
@@ -264,7 +264,7 @@ class TestPassManagerVerify:
         assert results == [("cancel_inverses", True), ("identity_pass", True)]
 
     def test_unitary_one_fails(self):
-        from qrisp.circuit.passes.cancel_inverses import cancel_inverses
+        from qrisp.circuit.pass_management.passes.cancel_inverses import cancel_inverses
 
         pm = PassManager([cancel_inverses, self._bad_unitary_pass()])
         qc = QuantumCircuit(2)
@@ -303,8 +303,8 @@ class TestPassManagerVerify:
         assert results == [("bad_unitary", True)]
 
     def test_unitary_returns_correct_length(self):
-        from qrisp.circuit.passes.cancel_inverses import cancel_inverses
-        from qrisp.circuit.passes.combine_single_qubit_gates import combine_single_qubit_gates
+        from qrisp.circuit.pass_management.passes.cancel_inverses import cancel_inverses
+        from qrisp.circuit.pass_management.passes.combine_single_qubit_gates import combine_single_qubit_gates
 
         pm = PassManager([cancel_inverses, combine_single_qubit_gates, identity_pass])
         qc = QuantumCircuit(2)
@@ -366,7 +366,7 @@ class TestPassManagerVerify:
 
     def test_visualize_culprits_only_bad_passes(self, capsys):
         """Only failing passes should be visualized."""
-        from qrisp.circuit.passes.cancel_inverses import cancel_inverses
+        from qrisp.circuit.pass_management.passes.cancel_inverses import cancel_inverses
 
         pm = PassManager([cancel_inverses, self._bad_unitary_pass()])
         qc = QuantumCircuit(2)
